feat/be/db test

See merge request s08-blockchain-contract-sub2/S08P22C208!95

+++ OMM/back/src/main/java/com/omm/admin/model/request/PunishMemberRequestDto.java
@@ -16,5 +16,5 @@ public class PunishMemberRequestDto {
     String type;
 
     @JsonProperty(value = "value")
-    int period;
+    Long period;
 }
